backend/requirement.txt
flask
numpy
mysql-connector-python
torch
torchvision
pillow
